Refactor panopticon registration
the `panopticon_registerer_for` was always used in conjunction with an
edition. It seems overcomplicated to repeat the stubbing of this
throughout our test suite.

Instead encapsulate that behaviour into a function
(`register_edition_with_panopticon`) which is unit tested
in one place and then we can use the simpler method elsewhere.